UI Updates, Formatting Corrections and minor edits
-Corrected formatting issues from initial conversion to Markdown.
-Lab 4, Exercise 1, Task 1, Step 5:
Added "Note: if you are signed in as a different user from a previous lab, first sign out, select the option to sign in as Other user and then sign in as Adatum\Beth."
-Lab 4, Exercise 1, Task 1, Step 7:
Added note: If you receive the error **"Verifying the challenge failed. Please refresh the page and try again."** this means the captcha feature is not currently working. This is a know issue that happens periodically with this site. If the issue does not resolve within 5 minutes you should skip to step 20 and continue with the lab.
-Lab 4, Exercise 1, Task 3, Step 18 (LAK) Step 17 (HL):
Added sentence. If DNS Manager is not on the task bar open Server Manager, select Tools and then select DNS.
-Lab 4, Exercise 1, Task 3, Step 21 (LAK) Step 21 (HL):
Added new step. "If the On-premises Account for Migration page appears, select **enter**, in the Office 365 dialog box type **Adatum\Administrator** with the password **Pa55w.rd**, select **ok** and then select **next**."
-Lab 4, Exercise 1, Task 3, Steps 28 and 29 (LAK) Steps 28 and 29 (HL):
Added two steps.
On the **Organization Configuration Transfer** page select **Review the objects before transferring from on-premises to exchange online** and select **next**.
On the **Organization Configuration Transfer - Review Objects** page, select the check box for **Objects** and **Organization Config**, and then select **next**.
Lab 04, Exercise 2, Step 14, Note:
removed the word "must"
Lab 04, Exercise 3, Task 1, Step 1:
Change to LON-CL1
Lab 04, Exercise 3, Task1,  Step 7 (HL) Step 8 (LAK):
Changed to On LON-EX1

Lab 04, Exercise 3, Task 2, Step 31:
changed last sentence to ...by running the follwoing commands,...