Notes added
Lab 04, Exercise 2, Step 25:
Added two notes:
Note: Not all users will appear in the same screen and you may not be able to scroll down the list with your mouse. If this happens, you can use the up and down arrows on the keyboard to scroll and the space bar to select users. Alternatively, you can assign licenses to the users individually or in smaller batches. It is important that all users listed are assigned a license in order for the users to access their migrated mailboxes.

Note: Not licensing these users now will create long delays in later labs.